move workspace folder into dashboard route
